check if file exists

+++ entrypoint.sh
@@ -41,19 +41,21 @@ for file in $FILES
 do
   case "$file" in
   "serverless/"*)
-    # get file name
-    filename=$(basename $file)
-    # get hash content of file
-    hash=$(cat $file)
-    # Use our dedicated profile and suppress verbose messages.
-    # All other flags are optional via `args:` directive.
-    aws s3 cp s3://${AWS_S3_BUCKET}/${hash}.yml ./ --profile push-s3-cfn
+    if test -f "$file"; then
+      # get file name
+      filename=$(basename $file)
+      # get hash content of file
+      hash=$(cat $file)
+      # Use our dedicated profile and suppress verbose messages.
+      # All other flags are optional via `args:` directive.
+      aws s3 cp s3://${AWS_S3_BUCKET}/${hash}.yml ./ --profile push-s3-cfn
 
-    aws cloudformation deploy --template-file ./${hash}.yml \
-        --stack-name $filename-${STAGE} \
-        --capabilities CAPABILITY_NAMED_IAM \
-        --parameter-overrides Stage=${STAGE} GitHash=${GITHASH} \
-        --profile push-s3-cfn
+      aws cloudformation deploy --template-file ./${hash}.yml \
+          --stack-name $filename-${STAGE} \
+          --capabilities CAPABILITY_NAMED_IAM \
+          --parameter-overrides Stage=${STAGE} GitHash=${GITHASH} \
+          --profile push-s3-cfn
+    fi
   ;;
   *       ) echo no ;;
   esac
